Them bao mat API

+++ service/index.php
@@ -8,6 +8,7 @@
 	require_once("Controllers/register.php");
 	require_once("Controllers/product.php");
 	require_once("Controllers/comment.php");
+	require_once("Controllers/token.php");
 	$app=new \Slim\App();
 
 	
@@ -30,30 +31,64 @@
 	});
 	$app->get("/carts/get",function($request,$response,$args)
 	{
-		$user_id=1;
-		$con=new Cart_Controller();
-		echo json_encode($con->getCartByUserId($user_id));
+
+		if($_SESSION["login"]==1)
+		{
+			$user_id=1;
+			$con=new Cart_Controller();
+			echo json_encode($con->getCartByUserId($user_id));
+		}
+		else
+		{
+			echo "Phải đăng nhập";
+		}
+		
 	});
-	$app->post("/carts/add",function($request,$response,$args)
+	$app->post("/carts/add",function($request,$response,$args)//sửa lại
 	{
-		$input=$request->getParsedBody();
-		$data= array(
+		$header=$request->getHeaderLine("Authorization");
+		$arr=Token_Controller::verifyJWT($header);
+		$login=new Login_Controller();
+		
+		if($_SESSION["login"]==1 && $_SESSION["admin"]==1 && $login->compareUser($arr))
+		{
+			$input=$request->getParsedBody();
+			$data= array(
 			"product_detail_id"=>$input["product_detail_id"],
 			"user_id"=>'1'
 			); 
-		$con=new Cart_Controller();
-		echo $con->addOneProductToCart($data);
+			$con=new Cart_Controller();
+			echo $con->addOneProductToCart($data);
+		}
+		else
+		{
+			echo "Không được phép !!!";
+		}
+		
+		
 	});
 	$app->get("/carts/delete/{product_detail_id}",function($request,$response,$args)
 	{
-        $product_detail_id=$args["id"];
-        die('123');
-		$data=array(
+		$header=$request->getHeaderLine("Authorization");
+		$arr=Token_Controller::verifyJWT($header);
+		$login=new Login_Controller();
+
+		if($_SESSION["login"]==1 && $_SESSION["admin"]==1 && $login->compareUser($arr))
+		{
+			$product_detail_id=$args["id"];
+      	
+			$data=array(
 			"user_id"=>'1',
 			"product_detail_id"=>$product_detail_id
 			);
-		$con=new Cart_Controller();
-		echo $con->deleteOrder($data);
+			$con=new Cart_Controller();
+			echo $con->deleteOrder($data);
+		}
+		else
+		{
+			echo "Không được phép !!!";
+		}
+  
 	});
 	$app->get("/comments/get/{id}",function($request,$response,$args)
 	{
@@ -63,31 +98,64 @@
 	});
 	$app->post("/comments/add",function($request,$response,$args)
 	{
-		$input=$request->getParsedBody();
-		$data=array(
+		if($_SESSION["login"]==1)
+		{
+			$input=$request->getParsedBody();
+			$data=array(
 			"product_detail_id"=>$input["product_detail_id"],
 			"user_id"=>$input["user_id"],
 			"content"=>$input["content"]
 			);
 
-		$con=new Comment_Controller();
-		echo $con->addComment($data);
+			$con=new Comment_Controller();
+			echo $con->addComment($data);
+		}
+		else
+		{
+			echo "Phải đăng nhập";
+		}
+		
 	});
 	$app->get("/orders/get",function($request,$response,$args)
 	{
-		$con=new Order_Controller();
-		echo $con->getAllOrder();
+		$header=$request->getHeaderLine("Authorization");
+		$arr=Token_Controller::verifyJWT($header);
+		$login=new Login_Controller();
+		if($_SESSION["login"]==1 && $_SESSION["admin"]==1 && $login->compareUser($arr))
+		{
+			$con=new Order_Controller();
+			echo $con->getAllOrder();
+		}
+		else
+		{
+			echo "Không được phép !!!";
+		}	
+		
 	});
 	$app->get("/orders/get/{id}",function($request,$response,$args)
 	{
-		$id=$args["id"];
-		$con=new Order_Controller();
-		echo $con->getOrderById($id);
+		$header=$request->getHeaderLine("Authorization");
+		$arr=Token_Controller::verifyJWT($header);
+		$login=new Login_Controller();
+
+		if($_SESSION["login"]==1 && $_SESSION["admin"]==1 && $login->compareUser($arr))
+		{
+			$id=$args["id"];
+			$con=new Order_Controller();
+			echo $con->getOrderById($id);
+		}
+		else
+		{
+			echo "Không được phép !!!";
+		}	
+		
 	});
 	$app->post("/orders/add",function($request,$response,$args)
 	{
-		$input=$request->getParsedBody();
-		$data=array(
+		if($_SESSION["login"]==1)
+		{
+			$input=$request->getParsedBody();
+			$data=array(
 			"user_id"=>'1',
 			"total_money"=>$input["total_money"],
 			"name"=>$input["name"],
@@ -96,14 +164,33 @@
 			"email"=>$input["email"],
 
 			);
-		$con=new Order_Controller();
-		echo $con->addOrder($data);
+			$con=new Order_Controller();
+			echo $con->addOrder($data);
+		}
+		else
+		{
+			echo "Không được phép !!!";
+		}	
+
+		
 	});
 	$app->get("/orders/delete/{id}",function($request,$response,$args)
 	{
-		$id=$args["id"];
-		$con=new Order_Controller();
-		echo $con->deleteOrder($id);
+		$header=$request->getHeaderLine("Authorization");
+		$arr=Token_Controller::verifyJWT($header);
+		$login=new Login_Controller();
+
+		if($_SESSION["login"]==1 && $_SESSION["admin"]==1 && $login->compareUser($arr))
+		{
+			$id=$args["id"];
+			$con=new Order_Controller();
+			echo $con->deleteOrder($id);
+		}
+		else
+		{
+			echo "Không được phép !!!";
+		}	
+		
 	});
 	$app->get("/products/get",function($request,$response,$args)
 	{
@@ -119,8 +206,14 @@
 	});
 	$app->post("/products/add",function($request,$response,$args)
 	{
-		$input=$request->getParsedBody();
-		$data=array(
+		$header=$request->getHeaderLine("Authorization");
+		$arr=Token_Controller::verifyJWT($header);
+		$login=new Login_Controller();
+
+		if($_SESSION["login"]==1 && $_SESSION["admin"]==1 && $login->compareUser($arr))
+		{
+			$input=$request->getParsedBody();
+			$data=array(
 			"name"=>$input["name"],
 			"description"=>$input["description"],
 			"type"=>$input["type"],
@@ -130,15 +223,35 @@
             "product_detail" => $input["product_detail"]
 			);
 
-		$con=new Product_Controller();
-		echo $con->addProduct($data);
+			$con=new Product_Controller();
+			echo $con->addProduct($data);
 
+		}
+		else
+		{
+			echo "Không được phép !!!";
+		}	
+
+		
 	});
 	$app->get("/products/delete/{id}",function($request,$response,$args)
 	{
-		$id=$args["id"];
-		$con=new Product_Controller();
-		echo $con->deleteProduct($id);
+		$header=$request->getHeaderLine("Authorization");
+		$arr=Token_Controller::verifyJWT($header);
+		$login=new Login_Controller();
+		if($_SESSION["login"]==1 && $_SESSION["admin"]==1 && $login->compareUser($arr))
+		{
+			$id=$args["id"];
+			$con=new Product_Controller();
+			echo $con->deleteProduct($id);
+
+		}
+		else
+		{
+			echo "Không được phép !!!";
+		}	
+
+		
 	});
 	$app->post("/register",function($request,$response,$args)
 	{
@@ -154,6 +267,7 @@
 	$app->get("/",function()
 	{
 		echo "<center><h1>Welcome to Shoe Shop !!!</h1></center>";
+
 	});
